pyproject.toml
[tool.poetry]
name = "facewatch"
version = "0.1.0"
description = "A facial recognition-based watchlist system"
authors = ["Lama Alosaimi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.3"
face-recognition = "^1.3.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"
flask-sqlalchemy = "^3.1.1"
pymongo = "^4.10.1"
flask-pymongo = "^2.3.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"